backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic-settings>=2.3.4
python-dotenv>=1.0.1
sqlmodel>=0.0.22
slowapi>=0.1.9
Pillow>=10.0.0
opencv-python-headless>=4.10.0.84
numpy>=1.26.0
python-magic>=0.4.27
torch>=2.1.0
torchvision>=0.16.0
basicsr>=1.4.2
realesrgan>=0.3.0
gfpgan>=1.3.8
redis>=5.0.7
rq>=1.16.2
pytest>=8.3.0
httpx>=0.27.0
